Add `__cxa_pure_virtual_` stub to resolve linker errors.
This is called by libstc++ when a pure virtual method is called by error. The
default implementation specifies a verbose error handler, which pulls in a bunch
of libc and blows up the executable size by ~40K. Replacing the noisy handler
with a simple loop prevents this and keeps the binary small.